requirements.txt
# Core dependencies
mlflow>=2.10.0
langchain>=0.1.0
langchain-community>=0.0.20
databricks-sdk>=0.18.0
databricks-agents>=0.1.0

# Data fetching
yfinance>=0.2.36
alpha-vantage>=2.3.1
pandas>=2.0.0
requests>=2.31.0

# Web scraping for unstructured data
beautifulsoup4>=4.12.0
newspaper3k>=0.2.8
feedparser>=6.0.10

# UI
gradio>=4.19.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
